\name{ findQ }

\alias{ findQ }

\title{ findQ }

\description{ Identifies the optimal inter-maxima quantile difference with which to set sample barcode-specific thresholds. Maximizes the pSinglet curve for quantiles 0.1-0.99. }

\usage{ findQ(res, extrema) }

\arguments{
  \item{res}{ Dataframe of pSinglet, pDoublet, and pNegative values across quantile sweep, as genereated by 'findThresh'. }
  \item{extrema}{ Vector of local pSinglet maxima values across quantile sweep, as generated by 'findThresh'. }
}

\value{ Quantile threshold applied to all barcode distributions during sample classificaiton. }

\author{ Chris McGinnis }

\examples{ round1.calls <- classifyCells(barTable, q=findQ(res_round1, extrema_round1)) }
